pyproject.toml
[project]
name = "telecom-cdr-auto-loader"
version = "1.0.0"
description = "ETL service"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
    "schedule>=1.2.2",
]

[dependency-groups]
dev = [
    "ruff>=0.4.0",
]
